+++ Python/tdw/vr_data/vive_button.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class ViveButton(Enum):
+    """
+    Enum values describing Vive controller buttons. Not all buttons on the controller are included here.
+    """
+
+    left_trackpad_click = 0
+    left_menu = 1